Moved BeamData to package
